Refactoring
moved the Reflector class into utilities
